@@ -1,10 +1,10 @@
 apply plugin: 'com.android.application'
 android {
     namespace 'org.alex.kitsune'
-    compileSdk 33
+    compileSdk 34
     defaultConfig {
         minSdkVersion 24
-        targetSdkVersion 33
+        targetSdkVersion 34
         versionCode 16
         versionName "1.8.1"
         vectorDrawables.useSupportLibrary = true
